ajuste arquitetura associado
