+++ ProjetoAssociados/Controllers/AssociadosController.cs
@@ -8,13 +8,13 @@ namespace ProjetoAssociados.Controllers
 {
     public class AssociadosController : Controller
     {
-        readonly private ApplicationDbContext _context;
+        //readonly private ApplicationDbContext _context;
         private readonly IEmpresaServices _empresaInterface;
         private readonly IAssociadoServices _associadoInterface;
 
-        public AssociadosController(ApplicationDbContext context, IEmpresaServices empresaInterface, IAssociadoServices associadoInterface)
+        public AssociadosController(IEmpresaServices empresaInterface, IAssociadoServices associadoInterface)
         {
-            _context = context;
+            //_context = context;
             _empresaInterface = empresaInterface;
             _associadoInterface = associadoInterface;
         }
@@ -22,7 +22,7 @@ namespace ProjetoAssociados.Controllers
         [HttpGet]
         public IActionResult Index()
         {
-            IEnumerable<AssociadoModel> associados = _associadoInterface.GetAssociados().Result; //_context.Associados;
+            IEnumerable<AssociadoModel> associados = _associadoInterface.GetAssociados().Result; 
             return View(associados);
         }
 
@@ -30,7 +30,7 @@ namespace ProjetoAssociados.Controllers
         public IActionResult Cadastrar()
         {
 
-            var AssociadosEmpresa = _empresaInterface.GetEmpresas().Result;  //_context.Empresas;
+            var AssociadosEmpresa = _empresaInterface.GetEmpresas().Result; 
 
             var empresaViewModel = new AssociadoViewModel();
             var checkboxListAssociados = new List<CheckBoxViewModel>();
@@ -50,30 +50,7 @@ namespace ProjetoAssociados.Controllers
         {
             if (ModelState.IsValid)
             {
-                var associado = new AssociadoModel();
-                associado.Nome = associadoViewModel.Nome;
-                associado.Cpf = associadoViewModel.Cpf;
-                associado.Empresas = new List<EmpresaModel>();
-
-                _context.Associados.Add(associado);
-                _context.SaveChanges();
-
-
-                foreach (var item in associadoViewModel.Empresas)
-                {
-                    if (item.Checked)
-                    {
-                        _context.AssociadosEmpresa.AddRange(new AssociadoModelEmpresaModel()
-                        {
-                            AssociadoId = associado.Id, //empresaViewModel.Id,
-                            EmpresaId = item.Id
-                        });
-
-                    }
-                }
-
-
-                _context.SaveChanges();
+                _associadoInterface.Cadastrar(associadoViewModel);
 
                 TempData["MensagemSucesso"] = "Cadastrado com sucesso!!";
 
@@ -88,44 +65,8 @@ namespace ProjetoAssociados.Controllers
         [HttpGet]
         public IActionResult Editar(int? Id)
         {
-            if (Id == null || Id == 0)
-            {
-                return NotFound();
-            }
 
-            AssociadoModel associado = _associadoInterface.GetAssociadoById(Id).Result; //_context.Associados.FirstOrDefault(x => x.Id == Id);
-
-            if (associado == null)
-            {
-                return NotFound();
-            }
-
-
-            var EmpresasAssociado = from c in _context.Empresas  //_empresaInterface.GetEmpresas().Result
-                                    select new
-                                    {
-                                        c.Id,
-                                        c.Nome,
-                                        Checked = ((from ce in _context.AssociadosEmpresa
-                                                    where (ce.AssociadoId == Id) & (ce.EmpresaId == c.Id)
-                                                    select ce).Count() > 0)
-                                    };
-
-            var associadoViewModel = new AssociadoViewModel();
-
-            associadoViewModel.Id = Id.Value;
-            associadoViewModel.Nome = associado.Nome;
-            associadoViewModel.Cpf = associado.Cpf;
-
-            var checkboxListAssociados = new List<CheckBoxViewModel>();
-
-            foreach (var item in EmpresasAssociado)
-            {
-                checkboxListAssociados.Add(new CheckBoxViewModel { Id = item.Id, Nome = item.Nome, Checked = item.Checked });
-            }
-
-            associadoViewModel.Empresas = checkboxListAssociados;
-
+            var associadoViewModel = _associadoInterface.GetEditar(Id).Result;
             return View(associadoViewModel);
 
         }
@@ -133,37 +74,11 @@ namespace ProjetoAssociados.Controllers
         [HttpPost]
         public IActionResult Editar(AssociadoViewModel associadoViewModel)
         {
+            AssociadoViewModel associadoEditada = new AssociadoViewModel();
+
             if (ModelState.IsValid)
             {
-                var associadoSelecionado = _associadoInterface.GetAssociadoById(associadoViewModel.Id).Result;   //_context.Associados.Find(associadoViewModel.Id);
-                associadoSelecionado.Nome = associadoViewModel.Nome;
-                associadoSelecionado.Cpf = associadoViewModel.Cpf;
-                //associadoSelecionado.DtNascimento = associadoViewModel.DtNascimento
-
-                var associadosEmpresa = _empresaInterface.GetEmpresasAssociado().Result;
-
-                foreach (var item in associadosEmpresa)
-                {
-                    if (item.AssociadoId == associadoViewModel.Id)
-                    {
-                        _context.Entry(item).State = Microsoft.EntityFrameworkCore.EntityState.Deleted;
-                    }
-                }
-
-                foreach (var item in associadoViewModel.Empresas)
-                {
-                    if (item.Checked)
-                    {
-                        _context.AssociadosEmpresa.Add(new AssociadoModelEmpresaModel()
-                        {
-                            EmpresaId = item.Id,
-                            AssociadoId = associadoSelecionado.Id
-                        });
-                    }
-                }
-
-
-                _context.SaveChanges();
+                associadoEditada = _associadoInterface.Editar(associadoViewModel).Result;
 
                 TempData["MensagemSucesso"] = "Editado com sucesso!!";
 
@@ -172,7 +87,7 @@ namespace ProjetoAssociados.Controllers
 
             TempData["MensagemErro"] = "Erro ao realizar edição!! Tente novamente";
 
-            return View(associadoViewModel);
+            return View(associadoEditada);
         }
 
         [HttpGet]
@@ -183,7 +98,7 @@ namespace ProjetoAssociados.Controllers
                 return NotFound();
             }
 
-            AssociadoModel associado = _associadoInterface.GetAssociadoById(Id).Result; //_context.Associados.FirstOrDefault(x => x.Id == Id);
+            AssociadoModel associado = _associadoInterface.GetAssociadoById(Id).Result; 
 
             if (associado == null)
             {
@@ -203,8 +118,6 @@ namespace ProjetoAssociados.Controllers
 
             _associadoInterface.DeleteAssociado(associadoModel.Id);            
 
-            //_context.Associados.Remove(associadoModel);
-            //_context.SaveChanges();
 
             return RedirectToAction("Index");
         }
