+++ ProjetoAssociados/Services/AssociadoServices/IAssociadoServices.cs
@@ -8,5 +8,9 @@ namespace ProjetoAssociados.Services.AssociadoServices
         Task<AssociadoModel> GetAssociadoById(int? id);
         void DeleteAssociado(int id);
 
+        Task<AssociadoViewModel> GetEditar(int? Id);
+        Task<AssociadoViewModel> Editar(AssociadoViewModel AssociadoViewModel);
+
+        void Cadastrar(AssociadoViewModel associadoViewModel);
     }
 }
